+++ src/app/static/src/tests/adr/actions.test.ts
@@ -88,7 +88,7 @@ describe("ADR actions", () => {
 
         await actions.getDatasets({commit, state, rootState} as any);
 
-        expect(commit.mock.calls.length).toBe(3);
+        expect(commit.mock.calls.length).toBe(4);
         expect(commit.mock.calls[0][0])
             .toStrictEqual({
                 type: ADRMutation.SetFetchingDatasets,
@@ -96,31 +96,46 @@ describe("ADR actions", () => {
             });
         expect(commit.mock.calls[1][0])
             .toStrictEqual({
+                type: ADRMutation.SetADRError,
+                payload: null
+            });
+        expect(commit.mock.calls[2][0])
+            .toStrictEqual({
                 type: ADRMutation.SetDatasets,
                 payload: [1]
             });
-        expect(commit.mock.calls[2][0])
+        expect(commit.mock.calls[3][0])
             .toStrictEqual({
                 type: ADRMutation.SetFetchingDatasets,
                 payload: false
             });
     });
 
-    it("resets datasets fetching if error response", async () => {
+    it("resets datasets and commits error fetching if error response", async () => {
         mockAxios.onGet(`/adr/datasets/`)
-            .reply(500, mockError("error"));
+            .reply(500, mockFailure("error"));
 
         const commit = jest.fn();
 
         await actions.getDatasets({commit, state, rootState} as any);
 
-        expect(commit.mock.calls.length).toBe(2);
+        expect(commit.mock.calls.length).toBe(4);
         expect(commit.mock.calls[0][0])
             .toStrictEqual({
                 type: ADRMutation.SetFetchingDatasets,
                 payload: true
             });
         expect(commit.mock.calls[1][0])
+            .toStrictEqual({
+                type: ADRMutation.SetADRError,
+                payload: null
+            });
+        expect(commit.mock.calls[2][0])
+            .toStrictEqual({
+                type: ADRMutation.SetADRError,
+                payload: mockError("error")
+            });
+        expect(commit.mock.calls[3][0])
             .toStrictEqual({
                 type: ADRMutation.SetFetchingDatasets,
                 payload: false
@@ -182,7 +197,10 @@ describe("ADR actions", () => {
         await actions.getUploadFiles({commit, state, rootState: root} as any);
 
         expect(commit.mock.calls[0][0].type).toBe("ADRError");
-        expect(commit.mock.calls[0][0].payload).toStrictEqual(mockError("test error"));
+        expect(commit.mock.calls[0][0].payload).toBeNull();
+
+        expect(commit.mock.calls[1][0].type).toBe("ADRError");
+        expect(commit.mock.calls[1][0].payload).toStrictEqual(mockError("test error"));
     });
 
     it("getUploadFiles gets dataset details and constructs upload files", async () => {
@@ -209,8 +227,11 @@ describe("ADR actions", () => {
 
         await actions.getUploadFiles({commit, state: adrState, rootState: root} as any);
 
-        expect(commit.mock.calls[0][0].type).toBe("SetUploadFiles");
-        expect(commit.mock.calls[0][0].payload).toStrictEqual({
+        expect(commit.mock.calls[0][0].type).toBe("ADRError");
+        expect(commit.mock.calls[0][0].payload).toBeNull();
+
+        expect(commit.mock.calls[1][0].type).toBe("SetUploadFiles");
+        expect(commit.mock.calls[1][0].payload).toStrictEqual({
             outputZip: {
                 index: 0,
                 displayName: "uploadFileOutputZip",
@@ -230,5 +251,110 @@ describe("ADR actions", () => {
                 resourceUrl: "http://test"
             }
         });
+    });
+
+    it("getUserCanUpload does nothing if no selected dataset", async () => {
+        const commit = jest.fn();
+        const root = mockRootState({
+            baseline: mockBaselineState()
+        });
+
+        await actions.getUserCanUpload({commit, state, rootState: root} as any);
+
+        expect(commit.mock.calls.length).toBe(0);
+        expect(mockAxios.history.get.length).toBe(0);
+    });
+
+    it("getUserCanUpload fetches updateable dataset and commits when user can upload", async () => {
+        const commit = jest.fn();
+        const root = mockRootState({
+            baseline: mockBaselineState({selectedDataset: {organization: {id: "test-org"}} as any})
+        });
+        const updateableOrgs = [{id: "other-org"}, {id: "test-org"}];
+        mockAxios.onGet(`adr/orgs?permission=update_dataset`)
+            .reply(200, mockSuccess(updateableOrgs));
+
+        await actions.getUserCanUpload({commit, state, rootState: root} as any);
+
+        expect(commit.mock.calls.length).toBe(1);
+        expect(commit.mock.calls[0][0]).toStrictEqual({type: "SetUserCanUpload", payload: true});
+    });
+
+    it("getUserCanUpload fetches updateable dataset and commits when user cannot upload", async () => {
+        const commit = jest.fn();
+        const root = mockRootState({
+            baseline: mockBaselineState({selectedDataset: {organization: {id: "test-org"}} as any})
+        });
+        const updateableOrgs = [{id: "other-org"}, {id: "other-org-2"}];
+        mockAxios.onGet(`adr/orgs?permission=update_dataset`)
+            .reply(200, mockSuccess(updateableOrgs));
+
+        await actions.getUserCanUpload({commit, state, rootState: root} as any);
+
+        expect(commit.mock.calls.length).toBe(1);
+        expect(commit.mock.calls[0][0]).toStrictEqual({type: "SetUserCanUpload", payload: false});
+    });
+
+    it("getUserCanUpload commits error", async () => {
+        const commit = jest.fn();
+        const root = mockRootState({
+            baseline: mockBaselineState({selectedDataset: {organization: {id: "test-org"}} as any})
+        });
+        mockAxios.onGet(`adr/orgs?permission=update_dataset`)
+            .reply(500, mockFailure("test-error"));
+
+        await actions.getUserCanUpload({commit, state, rootState: root} as any);
+        expect(commit.mock.calls.length).toBe(1);
+        expect(commit.mock.calls[0][0].type).toBe("ADRError");
+        expect(commit.mock.calls[0][0].payload).toStrictEqual(mockError("test-error"));
+    });
+
+    it("getUserCanUpload sets organisation on selectedDataset if necessary", async () => {
+        const commit = jest.fn();
+        const root = mockRootState({
+            baseline: mockBaselineState({selectedDataset: {id: "test-dataset"}} as any)
+        });
+        const adr = mockADRState({
+            datasets: [{id: "test-dataset", resources: [], organization: {id: "test-org"}}],
+            schemas: {baseUrl: "http://test"} as any
+        });
+
+        mockAxios.onGet(`adr/orgs?permission=update_dataset`)
+            .reply(200, mockSuccess([{id: "test-org"}]));
+
+        await actions.getUserCanUpload({commit, state: adr, rootState: root} as any);
+
+        expect(commit.mock.calls.length).toBe(2);
+        expect(commit.mock.calls[0][0]).toBe("baseline/SetDataset");
+        expect(commit.mock.calls[0][1].id).toBe("test-dataset");
+        expect(commit.mock.calls[0][1].organization).toStrictEqual({id: "test-org"});
+        expect(commit.mock.calls[0][2]).toStrictEqual({root: true});
+        expect(commit.mock.calls[1][0]).toStrictEqual({type: "SetUserCanUpload", payload: true});
+    });
+
+    it("getUserCanUpload fetches dataset metadata to get organization if necessary", async () => {
+        const commit = jest.fn();
+        const root = mockRootState({
+            baseline: mockBaselineState({selectedDataset: {id: "test-dataset"}} as any)
+        });
+        const adr = mockADRState({
+            datasets: [],
+            schemas: {baseUrl: "http://test"} as any
+        });
+
+        const datasetResponse = {id: "test-dataset", resources: [], organization: {id: "test-org"}}
+        mockAxios.onGet(`adr/datasets/test-dataset`)
+            .reply(200, mockSuccess(datasetResponse));
+        mockAxios.onGet(`adr/orgs?permission=update_dataset`)
+            .reply(200, mockSuccess([{id: "test-org"}]));
+
+        await actions.getUserCanUpload({commit, state: adr, rootState: root} as any);
+
+        expect(commit.mock.calls.length).toBe(2);
+        expect(commit.mock.calls[0][0]).toBe("baseline/SetDataset");
+        expect(commit.mock.calls[0][1].id).toBe("test-dataset");
+        expect(commit.mock.calls[0][1].organization).toStrictEqual({id: "test-org"});
+        expect(commit.mock.calls[0][2]).toStrictEqual({root: true});
+        expect(commit.mock.calls[1][0]).toStrictEqual({type: "SetUserCanUpload", payload: true});
     });
 });