+++ NEWS.md
@@ -1,6 +1,18 @@
-# hint 1.22.0
+# hint 1.25.0
 
 * Creates consistency on how the enter button works
+
+# hint 1.24.0
+
+* Make login email case insensitive
+
+# hint 1.23.0
+
+* Improved handling of ADR timeouts
+
+# hint 1.22.0
+
+* Check if user has ADR upload permission
 
 # hint 1.21.0